requirements.txt
aiohttp==3.8.1
aioredis==2.1.0
discord.py>=2.2.0
Pillow>=9.5.0